# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ujima SACCO AI Pride Ecosystem API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new applicant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: New token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/password-reset-request:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequestInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/password-reset:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /auth/mfa/verify:
    post:
      operationId: verifyMfa
      tags: [auth]
      summary: Verify MFA token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaVerifyInput"
      responses:
        "200":
          description: MFA verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # ── USERS ─────────────────────────────────────────────────────────────────
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for current member
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── AI ASSESSMENT ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH SCHEMAS ────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [firstName, lastName, email, password, phone, idNumber]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        idNumber:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        mfaToken:
          type: ["string", "null"]
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    PasswordResetRequestInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    PasswordResetInput:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
    MfaVerifyInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
        sessionToken:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: ["string", "null"]
        requiresMfa:
          type: ["boolean", "null"]
        sessionToken:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [applicant, admin]
        isActive:
          type: boolean
        mfaEnabled:
          type: ["boolean", "null"]
        createdAt:
          type: string
...
    DocumentInput:
      type: object
      required: [fileName, fileType, documentType, fileData]
      properties:
        fileName:
          type: string
        fileType:
          type: string
        documentType:
          type: string
          enum: [national_id, passport, kra_pin, payslip, bank_statement, business_certificate, title_deed, other]
        applicationId:
          type: ["integer", "null"]
        fileData:
          type: string
          description: Base64 encoded file content
    Document:
      type: object
      required: [id, userId, fileName, documentType, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        applicationId:
          type: ["integer", "null"]
        fileName:
          type: string
        fileType:
          type: string
        documentType:
          type: string
        status:
          type: string
          enum: [uploaded, verified, rejected]
        createdAt:
          type: string

    # ── AI SCHEMAS ──────────────────────────────────────────────────────────
...